moved things into component folder
