+++ app/src/features/surveys/view/summary-results/components/FileSummaryResults.tsx
