+++ app/src/features/surveys/view/summary-results/components/NoSummaryResults.tsx
