+++ app/src/features/surveys/view/summary-results/components/SummaryResultsLoading.tsx
